Fix datatypes for doorbell
The doorbell is an uint32_t so the current doorbell needs to be stored
as uint32_t too. The iptsd_control_doorbell function needs to be able to
return either a uint32_t (valid doorbell), or a negative value (error),
so its return type needs to be int64_t.

+++ control.c
@@ -80,7 +80,7 @@ int iptsd_control_flush(struct iptsd_control *control)
 	return 0;
 }
 
-uint32_t iptsd_control_doorbell(struct iptsd_control *control)
+int64_t iptsd_control_doorbell(struct iptsd_control *control)
 {
 	iptsd_control_wait_for_device(control);
